Move TS tests out of not-supported directory
