+++ test/typescript/policy-tests.ts
@@ -1,4 +1,4 @@
-import { policies } from "../../../main";
+import { policies } from "../../main";
 import LoadBalancingPolicy = policies.loadBalancing.LoadBalancingPolicy;
 import TokenAwarePolicy = policies.loadBalancing.TokenAwarePolicy;
 import ReconnectionPolicy = policies.reconnection.ReconnectionPolicy;
